added Badges API integration tests
