+++ src/server.js
@@ -1,49 +1,8 @@
-const express = require('express')
-const path = require('path')
-const responseTime = require('response-time')
-const badgesAPIRoutes = require('./api/badges')
 const dotenv = require('dotenv')
-const bodyParser = require('body-parser')
-
 dotenv.config()
+const app = require('./app')
 
-//Init express
-const app = express()
-const port = 3000
-
-// -----------------------------------------------
-
-// Contenido estático
-app.use(
-    '/comunidad',
-    express.static(path.join(__dirname, '..', 'html', 'comunidad.html'))
-)
-app.use(
-    '/diario',
-    express.static(path.join(__dirname, '..', 'html', 'diario.html'))
-)
-
-// Loader.io
-app.use('/loaderio-*', (req, res) => {
-    res.send(process.env.LOADERIO_VERIFICATION_TOKEN)
-})
-
-app.use(bodyParser.urlencoded({ extended: true }))
-
-// Para ver el rendimiento de las peticiones/respuestas
-app.use(responseTime((req, res, time) => console.log(req.path, time)))
-
-// API: Insignias de todos los usuarios
-app.use('/badges', badgesAPIRoutes)
-
-// Si llegamos aqui es que la petición no se ha tratado
-app.use(function (req, res) {
-    res.status(404).json({
-        error: 'Aquí no hay nada de nada...sigue buscando...',
-    })
-})
-
-// -----------------------------------------------
+const port = process.env.PORT || 3000
 
 // Start Server
 app.listen(port, () => {
